requirements.txt
opencv-python
sqlalchemy
psycopg2
kafka-python